+++ src/modules/common/Quote.tsx
