+++ src/modules/common/cards/BookPreviewCard.tsx
