+++ src/modules/common/cards/InterviewPreviewCard.tsx
@@ -1,23 +1,29 @@
 import React from 'react';
 import { Heading, Flex, Box, Text, Image, Tag } from '@chakra-ui/react';
-import { CLOUDINARY_URL } from '../../constants';
+import { CLOUDINARY_URL } from '../../../constants';
 
-const ArticlePreviewCard = ({
+const InterviewPreviewCard = ({
   slug,
   title,
-  imageUniqueIdentifier,
   description,
   readTime,
+  imageUniqueIdentifier,
 }) => {
   return (
-    <Box key={slug} w="100%" shadow="sm" p={4} borderBottom="5px solid #00A3C4">
+    <Box
+      key={imageUniqueIdentifier}
+      w="100%"
+      shadow="sm"
+      borderBottom="5px solid #00A3C4"
+      p={4}
+    >
       <Flex
         flexDirection={['column', 'row']}
         alignItems="center"
         justifyContent="center"
       >
         <Image
-          src={`${CLOUDINARY_URL}c_scale,h_205,w_300/images/articles/${slug}_${imageUniqueIdentifier}.jpg`}
+          src={`${CLOUDINARY_URL}c_scale,h_205,w_300/images/interviews/${slug}_${imageUniqueIdentifier}.jpg`}
           alt={title}
           w={['100%', 300]}
           mr={[0, 0, 4]}
@@ -29,8 +35,8 @@ const ArticlePreviewCard = ({
           </Heading>
           <Text mt={2}>{description}</Text>
           <Box w="100%" textAlign={['center', 'left']}>
-            <Tag colorScheme="cyan" size="md" mt={2} width="116px">
-              Read Article
+            <Tag colorScheme="cyan" size="md" mt={2} width="130px">
+              Read Interview
             </Tag>
             <Text fontStyle="italic" color="grey" mt={1}>
               {`${readTime.text}`}
@@ -42,4 +48,4 @@ const ArticlePreviewCard = ({
   );
 };
 
-export default ArticlePreviewCard;
+export default InterviewPreviewCard;
