+++ src/modules/readmore/ReadMoreBookSummaries.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { Heading, Flex, Box, Link, Stack } from '@chakra-ui/react';
 // @ts-ignore
-import { frontMatter as bookSummariesPosts } from '../../pages/book-summaries/**/*.mdx';
+import { frontMatter as bookSummariesPosts } from '@/pages/book-summaries/**/*.mdx';
 import NextLink from 'next/link';
-import { FrontMatterBookSummaries } from '../../types/types';
-import BookPreviewCard from '../cards/BookPreviewCard';
+import { FrontMatterBookSummaries } from '@/types/types';
+import BookPreviewCard from '@/modules/common/cards/BookPreviewCard';
 
 interface Props {
   category: string;
