+++ src/pages/articles/[slug].tsx
@@ -1,24 +1,24 @@
 import React from 'react';
-import PageContainer from '../../containers/PageContainer';
+import PageContainer from '@/containers/PageContainer';
 import { Heading, Box, Flex, Text } from '@chakra-ui/react';
-import PageSeo from '../../components/partials/PageSeo';
+import PageSeo from '@/modules/common/PageSeo';
 import dynamic from 'next/dynamic';
-import { FrontMatterArticles } from '../../types/types';
-import EmailSubscription from '../../components/email/EmailSubscription';
-import { AmazonAdvert } from '../../components/adverts/AmazonAdvert';
-import { Image960x660 } from '../../components/image_components/Image960x660';
-import { AmazonCta } from '../../components/partials/AmazonCta';
-import { Image266x400 } from '../../components/image_components/Image266x400';
+import { FrontMatterArticles } from '@/types/types';
+import EmailSubscriptionForm from '@/modules/email/EmailSubscriptionForm';
+import { AmazonAdvert } from '@/modules/adverts/AmazonAdvert';
+import { Image960x660 } from '@/modules/common/images/Image960x660';
+import { AmazonCta } from '@/modules/common/AmazonCta';
+import { Image266x400 } from '@/modules/common/images/Image266x400';
 import readingTime from 'reading-time';
 
 // Remote packages
 import fs from 'fs';
 import matter from 'gray-matter';
 import path from 'path';
-import { articleFilePaths, ARTICLE_PATH } from '../../utils/mdxUtils';
+import { articleFilePaths, ARTICLE_PATH } from '@/utils/mdxUtils';
 import renderToString from 'next-mdx-remote/render-to-string';
 import hydrate from 'next-mdx-remote/hydrate';
-import WrittenBy from '@/modules/common/components/WrittenBy';
+import WrittenBy from '@/modules/common/WrittenBy';
 
 type Props = {
   frontMatter: FrontMatterArticles;
@@ -80,7 +80,7 @@ export default function InterviewLayout({ frontMatter, source }: Props) {
         {content}
       </Box>
       <AmazonAdvert />
-      <EmailSubscription />
+      <EmailSubscriptionForm />
     </PageContainer>
   );
 }
